shot_scale_model/requirements.txt
# MTC API Utils package
git+https://github.com/mediatechnologycenter/api-utils.git

Pillow~=9.3.0
torch~=1.13.1
torchvision~=0.14.1
huggingface_hub
